requirements.txt
fastapi
python-dotenv
tortoise-orm[asyncpg]
pycryptodome
uvicorn